Scene Change, UI Buttons, Volume Slider
All of this in one commit because they are connected, added a SceneManager prefab with application handler which plays a scene based on build profile index, it is called when pressing play. Or quit then we quit the application.
Then in the main menu scene is there UI buttons for Play and Quit and also volume sliders which control the audio of the Master, Music and SFX mixer groups.
The main menu scene is the start scene. The buttons and sliders are prefabs aswell